move debug print function to Debug file
