+++ Fallout4VR_Body.vcxproj
@@ -175,6 +175,7 @@ copy "$(SolutionDir)x64\Release\FRIK.pdb" "H:\mo2\fo4vr\mods\FRIK_69\F4SE\Plugin
     <ClCompile Include="BSFlattenedBoneTree.cpp" />
     <ClCompile Include="Config.cpp" />
     <ClCompile Include="ConfigurationMode.cpp" />
+    <ClCompile Include="Debug.cpp" />
     <ClCompile Include="F4VRBody.cpp" />
     <ClCompile Include="GunReload.cpp" />
     <ClCompile Include="HandPose.cpp" />
@@ -200,6 +201,7 @@ copy "$(SolutionDir)x64\Release\FRIK.pdb" "H:\mo2\fo4vr\mods\FRIK_69\F4SE\Plugin
     <ClInclude Include="BSFlattenedBoneTree.h" />
     <ClInclude Include="Config.h" />
     <ClInclude Include="ConfigurationMode.h" />
+    <ClInclude Include="Debug.h" />
     <ClInclude Include="F4VRBody.h" />
     <ClInclude Include="GunReload.h" />
     <ClInclude Include="HandPose.h" />
