+++ GunReload.cpp
@@ -4,39 +4,12 @@
 #include "VR.h"
 #include "Offsets.h"
 #include "MiscStructs.h"
+#include "Debug.h"
 
 namespace F4VRBody {
 	GunReload* g_gunReloadSystem = nullptr;
 
 	float g_animDeltaTime = -1.0f;
-
-
-	void printNodes(NiNode* nde, long long curTime) {
-		_MESSAGE("%d %s : children = %d %d: local %f %f %f %f %f %f %f %f %f %f %f %f %f %f %f", curTime, nde->m_name.c_str(), nde->m_children.m_emptyRunStart, (nde->flags & 0x1),
-			nde->m_localTransform.rot.arr[0],
-			nde->m_localTransform.rot.arr[1],
-			nde->m_localTransform.rot.arr[2],
-			nde->m_localTransform.rot.arr[3],
-			nde->m_localTransform.rot.arr[4],
-			nde->m_localTransform.rot.arr[5],
-			nde->m_localTransform.rot.arr[6],
-			nde->m_localTransform.rot.arr[7],
-			nde->m_localTransform.rot.arr[8],
-			nde->m_localTransform.rot.arr[9],
-			nde->m_localTransform.rot.arr[10],
-			nde->m_localTransform.rot.arr[11],
-			nde->m_localTransform.pos.x, nde->m_localTransform.pos.y, nde->m_localTransform.pos.z);
-
-
-		if (nde->GetAsNiNode()) {
-			for (auto i = 0; i < nde->m_children.m_emptyRunStart; ++i) {
-				auto nextNode = nde->m_children.m_data[i] ? nde->m_children.m_data[i]->GetAsNiNode() : nullptr;
-				if (nextNode) {
-					printNodes((NiNode*)nextNode, curTime);
-				}
-			}
-		}
-	}
 
 	void GunReload::DoAnimationCapture() {
 		if (!startAnimCap) {
@@ -54,7 +27,7 @@ namespace F4VRBody {
 		}
 
 		NiNode* weap = getChildNode("Weapon", (*g_player)->firstPersonSkeleton);
-	//	printNodes(weap, elapsed);
+		//printNodes(weap, elapsed);
 	}
 
 	bool GunReload::StartReloading() {
