+++ Skeleton.cpp
@@ -16,6 +16,7 @@ extern PapyrusVRAPI* g_papyrusvr;
 extern OpenVRHookManagerAPI* _vrhook;
 
 using namespace std::chrono;
+
 namespace F4VRBody {
 
 	float defaultCameraHeight = 120.4828;
@@ -90,48 +91,6 @@ namespace F4VRBody {
 		return boneTreeMap[boneName];
 	}
 	
-	void Skeleton::printChildren(NiNode* child, std::string padding) {
-		padding += "....";
-		_MESSAGE("%s%s : children = %d hidden: %d: local (%2f, %2f, %2f) world (%2f, %2f, %2f)", padding.c_str(), child->m_name.c_str(), child->m_children.m_emptyRunStart, (child->flags & 0x1),
-			child->m_localTransform.pos.x,
-			child->m_localTransform.pos.y,
-			child->m_localTransform.pos.z,
-			child->m_worldTransform.pos.x,
-			child->m_worldTransform.pos.y,
-			child->m_worldTransform.pos.z);
-
-		//_MESSAGE("%s%s : children = %d : worldbound %f %f %f %f", padding.c_str(), child->m_name.c_str(), child->m_children.m_emptyRunStart,
-		//	child->m_worldBound.m_kCenter.x, child->m_worldBound.m_kCenter.y, child->m_worldBound.m_kCenter.z, child->m_worldBound.m_fRadius);
-
-		if (child->GetAsNiNode())
-		{
-			for (auto i = 0; i < child->m_children.m_emptyRunStart; ++i) {
-				//auto nextNode = child->m_children.m_data[i] ? child->m_children.m_data[i]->GetAsNiNode() : nullptr;
-				auto nextNode = child->m_children.m_data[i];
-				if (nextNode) {
-					this->printChildren((NiNode*)nextNode, padding);
-				}
-			}
-		}
-	}
-
-
-	void Skeleton::printNodes(NiNode* nde) {
-		// print root node info first
-		_MESSAGE("%s : children = %d hidden: %d: local (%f, %f, %f)", nde->m_name.c_str(), nde->m_children.m_emptyRunStart, (nde->flags & 0x1),
-			nde->m_localTransform.pos.x,nde->m_localTransform.pos.y,nde->m_localTransform.pos.z);
-
-		std::string padding = "";
-
-		for (auto i = 0; i < nde->m_children.m_emptyRunStart; ++i) {
-		//	auto nextNode = nde->m_children.m_data[i] ? nde->m_children.m_data[i]->GetAsNiNode() : nullptr;
-			auto nextNode = nde->m_children.m_data[i];
-			if (nextNode) {
-				this->printChildren((NiNode*)nextNode, padding);
-			}
-		}
-	}
-
     void Skeleton::rotateWorld(NiNode *nde) {
         Matrix44 mat;
         mat.data[0][0] = -1.0;
@@ -578,14 +537,6 @@ namespace F4VRBody {
 
 		initBoneTreeMap(_root);
 		return true;
-	}
-
-	void Skeleton::positionDiff() {
-		NiPoint3 firstpos = _playerNodes->HmdNode->m_worldTransform.pos;
-		NiPoint3 skellypos = _root->m_worldTransform.pos;
-
-		_MESSAGE("difference = %f %f %f", (firstpos.x - skellypos.x), (firstpos.y - skellypos.y), (firstpos.z - skellypos.z));
-
 	}
 
 	NiPoint3 Skeleton::getPosition() {
